Testing out the paywall
I think it'd be cool to see if we can test out this Lightning Paywall for a start up. We can publish the notes we create when we collaborate for the Socratic Seminar and place it behind the paywall so it's not just links to documents but in depth information about topics. The goal of this wouldn't be to generate revenue but to see if we can get some traction and test the market for them a bit.

In order to implement the paywall, it looks like all we have to do is copy and paste this script onto a new page. But I'm not sure how  the website will process this PR because it looks like it takes markdown and translates it into HTML/CSS for the site. 

Can you take a look at this? 

You can also check out what it looks like on this codepen.
https://codepen.io/ecurrencyhodler/pen/ZEQPXMW